requirements.txt
astroid==4.0.2
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
dill==0.4.0
discord-interactions==0.4.0
dotenv==0.9.9
Flask==3.1.2
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
platformdirs==4.5.0
pycparser==2.23
pylint==4.0.4
PyNaCl==1.6.1
python-dotenv==1.2.1
ruff==0.14.7
tomlkit==0.13.3
Werkzeug==3.1.4
